Add MapBox map to show page
In show.ejs
Add script to stringify campground object.

This will be used by our showPageMap script file which will parse the coordinates to place a marker and center the user on the campground location

In showPageMap.js
Parse camp from stringified JSON.
center at the coordinates and place a marker at coordinates.